Added ARCOS functionality, fixed PlotWindow displaying errors, made canvas resizable
Canvas now displays in a separate resizable window locked to the
aspect ratio of the world.

New Plotwindow requires only a one line change to work with SoaR's
UI: An argument specifying what Toplevel() function to call when
creating the window. This can then be passed from SoarUI->client->
sim->brain, which then passes it to the PlotWindow constructor

ARCOS Functionality: Can connect to, disconnect from, and send
various commands (most of the necessary ones) to Pioneer robots